+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+* The UUID is no longer required when adding a dependency.
 * If the project is explicitly specified, it is considered "shared" and existing
   dependencies are never removed.